backend/requirements.txt
django>=4.2,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
requests>=2.31
python-dateutil>=2.8